Changed selecting area of characters
No need for pressing control anymore.
Selecting a character or area of characters is now handled by
the same function. If the selected area covers only one character,
it's selected if it's not selected yet and unselected if it's
already selected.
If the selected area covers many characters, all those characters
are selected.