Redesigned find
Find was implemented previously to find only the next match saving
index to continue search from that. It was complex and searching
backwards was difficult to implement.

Now, paths to all matching items are saved first and using an index,
next and previous is easy to implement(finding the actual widget if
path is to a character is ugly though).